Implemented notification interval handling
Rather than always send a notification whenever there's a change to the
battery status of the device, the notification system should keep track
of when the last notification was, and only send a new notification if
the elapsed time period exceeds the preferred interval.

This involved adding the new interval settings to the preferences
activity, as well as improved behaviors in the notification module to
check and reset the intervals appropriately.